requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydub==0.25.1
requests==2.32.3
openai==1.37.1
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.2
